requirements.txt
requests>=2.28,<3.0
python-dotenv>=1.0,<2.0
python-dateutil>=2.8,<3.0
pillow>=10.0,<12.0
streamlit>=1.30,<2.0
